+++ src/app/navscroll/navscroll-examples/navscroll-examples.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'it-navscroll-examples',
+  templateUrl: './navscroll-examples.component.html',
+  styleUrl: './navscroll-examples.component.scss',
+})
+export class NavscrollExamplesComponent {}